+++ src/icloudpd/status.py
@@ -4,17 +4,20 @@ from typing import Optional
 
 
 class Status(Enum):
-    NOT_NEED_MFA = "not_need_mfa"
+    NO_INPUT_NEEDED = "no_input_needed"
     NEED_MFA = "need_mfa"
     SUPPLIED_MFA = "supplied_mfa"
     CHECKING_MFA = "checking_mfa"
+    NEED_PASSWORD = "need_password"
+    SUPPLIED_PASSWORD = "supplied_password"
+    CHECKING_PASSWORD = "checking_password"
 
 
 class StatusExchange:
     def __init__(self) -> None:
         self.lock = Lock()
-        self._status = Status.NOT_NEED_MFA
-        self._code: Optional[str] = None
+        self._status = Status.NO_INPUT_NEEDED
+        self._payload: Optional[str] = None
 
     def get_status(self) -> Status:
         with self.lock:
@@ -28,18 +31,18 @@ class StatusExchange:
             else:
                 return False
 
-    def set_code(self, code: str) -> bool:
+    def set_payload(self, payload: str) -> bool:
         with self.lock:
-            if self._status != Status.NEED_MFA:
+            if self._status != Status.NEED_MFA and self._status != Status.NEED_PASSWORD:
                 return False
 
-            self._code = code
-            self._status = Status.SUPPLIED_MFA
+            self._payload = payload
+            self._status = Status.SUPPLIED_MFA if self._status == Status.NEED_MFA else Status.SUPPLIED_PASSWORD
             return True
 
-    def get_code(self) -> Optional[str]:
+    def get_payload(self) -> Optional[str]:
         with self.lock:
-            if self._status not in [Status.SUPPLIED_MFA, Status.CHECKING_MFA]:
+            if self._status not in [Status.SUPPLIED_MFA,Status.CHECKING_MFA, Status.SUPPLIED_PASSWORD,Status.CHECKING_PASSWORD]:
                 return None
 
-            return self._code
+            return self._payload